refactor: 부가 기능 관련 폴더/파일명 private -> extra로 변경
